Dump threads if the task doesn't progress
Summary:
This adds a part of the logic that dumps a few stack traces
inside of the task if it stops reporting progress.

This will dump 5 thread dumps evenly distributed in time after the
progress has not been sent for the half the timeout period.

For example with our default timeout of 10 minutes once we did not
heartbeat for 5 minutes we will dump one thread dump a minute.

Test Plan:
creating a special job that sits in the map() for a long time without
reporting progress.

Reviewers: rvadali, aching

Reviewed By: rvadali