Introducing a timeout for umbilical
Summary:
This introduces a timeout on umbilical protocol proxy in the
child process.
The problem we are seeing is the status updates getting stuck.
Introducing a timeout will help because we will be able to retry the
call instead of the task just dying 600 seconds into getting stuck.

Test Plan: ant test, mrtestcorona and shadow/snapshot

Reviewers: aching, rvadali

Reviewed By: aching

Task ID: 1439826