Fix timeout handling
Summary:
We weren't getting a SocketTimoutException in the error
handling code, instead we were getting an IOException caused by the
SocketTimeout exception.
I am changing the code such that only non IOExceptions are wrapped and
the IOExceptions are just rethrown. This way we will get the
sockettimeout not modified.

Test Plan: unittest and then live

Reviewers: aching, rvadali

Reviewed By: aching

Task ID: 1439826