Modification of Hadoop codebase to allow corona task tracker to utilize a list of working directories from a different file
Summary: Modified codebase to allow Corona to know which directories are good to write and which is bad using a list of mounts by using DataDirReader. It looks at the config path specified by the new parameter in mapred-default.xml called mapred.localdir.confpath.

Test Plan:
Compiles fine and will test this on our test cluster to see if it successfully utilizes different list of directories.

This code is based on code that is going to be used in our datanodes

Reviewers: dms, pknowles

Reviewed By: dms