Add aliases to the list of federation keys.
Summary:
We had introduced a fs.default.name.aliases key so that we can
deal with aliases for a cluster during avatar failover. Although the
problem with this currently in federated clusters is that both NameNodes
share the same configuration and hence incorrectly share the same set of
aliases which is dangerous.

We need service level aliases such as fs.default.name.aliases.DFSTEST.

Test Plan: Tested with DFSTEST zk.

Reviewers: weiyan, tomasz

Reviewed By: tomasz

Task ID: 1450996