Dump threads when CJT doesn't heartbeat
Summary:
When CJT doesn't heartbeat to the parent for a while we can
dump all its threads into the log so later we can figure out what was
happening. This can help us with diagnosing the issue of the tasks not
being able to talk to the CJT.
Two things about this diff though:
1) if you want to I can make the thresholds configurable
2) it might make sense to encapsulate the functionality of "dump threads
if there's no update for a while" in a separate class and just use it
from here since it seems to be a common pattern.

Test Plan: unittests

Reviewers: rvadali, aching

Reviewed By: rvadali