Increaseing zookeeper connect and session timeout for unit tests.
Summary:
As per Thawan's suggestion from the zeus team the best thing
to do to avoid the "Timed out connecting to zookeeper" exceptions we see
in our unit tests is to use a higher sesstion timeout and connect
timeout.

Test Plan:
1) Ran a single avatar unit test 100 times using mrunit and
verified that it never timed out connecting to zookeeper.

Reviewers: tomasz, hkuang

Reviewed By: tomasz