Log the CTT exception better
Summary:
We have seen the CTT hanging because the configuration was not
complete. The exception was written into the .out file and the process
was alive because some threads stuck around.

This will put the error into the log file instead plus will make sure
the process goes down.

Test Plan: build and run

Reviewers: rvadali, aching

Reviewed By: rvadali

Task ID: 1463320