[L38-24-tests] Clean up checkpointing in tests
Summary:
-Since we assert transaction ids, I want to make sure that after the
cluster startup in tests, the first checkpoint is done, to account for
it in the txid.Of course, we can enforce no waiting for checkpoint if
needed.

Otherwise it interferes with the manual checkpoints and messes up the
txids.

-Also, for MiniAvatarCluster, I set the sleep for standby to be shorter,
as for the tests with multiple checkpoints it extends them
significantly.

Test Plan: ant test for HA

Reviewers: hkuang, avf, pritam

Reviewed By: avf