Randomly choose ports instead of sequentially.
Summary:
When running two unit tests on the same machine, since they end
up choosing the same ports this causes the test to start later to get
stalled considerably since it tries to connect to multiple ports in a
sequence before it finds the first port that it can connect to.

Test Plan:
All unit tests.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: tomasz, hkuang, weiyan, sdong

Reviewed By: hkuang