[L38-25-ckpt] Add logic for readin fs.checkpoint.size
Summary:
This patch reads old configuration key fs.checkpoint.size if the new one
is not present. I checked that on dfs1, average transaction count is
below 100bytes (~90 or so), so I think it's a fair and round assumption
to make that the transaction is about 100 bytes long.

Hence, if the old key is present the transaction count for checkpoint is
inferred by dividing the isze over 100.

I also added the fs.checkpoint.txns to core-default instead of
fs.checkpoint.size

Test Plan: added testcase

Reviewers: pritam, hkuang, avf

Reviewed By: pritam