Add a HighMemoryUsageKilled counter for a Hadoop job
Summary:
Added the counter to calculate the number of tasks which have been terminated
because of high memory by the TaskMemoryManagerThread. Using the "high-memory"
keyword in the diagnostic message to identify such failed/killed tasks.

Test Plan: Not sure how to test it. A preliminary version to know if, on right path.

Reviewers: rvadali, kevinwilfong, sambavim

Reviewed By: rvadali

CC: mukundn, dms

Task ID: 1444325