Lost task tracker should not fail succeeded map tasks in map-only job
Summary:
For map-only jobs, the map tasks write output directly to HDFS. So there
is no need to run successful map tasks when the task tracker is lost.

Test Plan:
Ran a single node sleep job with many sleep tasks. Killed the task
tracker and verified that the successful map tasks are not killed.

Reviewers: dms, aching

Reviewed By: aching

Task ID: 1733245