HDFS DataNode: avoid two string concatenation when ClientTraceLog is not enabled
Summary:
Currently, we always calculate localAddress and remoteAddress even if ClientTraceLog is not enabled. Both of them are string concatenations. Remove them when ClientTraceLog is not enabled. Another place uses remoteAddress is thread name. Replace that with just addresses without ports. I think that's good enough for thread name.

The performance improvement is minor but it won't be worse.

Test Plan: N/A

Reviewers: weiyan, hkuang

Reviewed By: hkuang