Fix problem with upgrade
Summary:
I found that when upgrading old image to the new layout, if the edits
are empty, we have conflict with the loaded and saved image digest,
since the saved image is different (e.g., layout version). During
upgrade we need to clean the digest, so it doesn't conflict.

Test Plan: Manually

Reviewers: pritam, hkuang

Reviewed By: hkuang