Enhance image/edits config validator for supporting uris
Summary:
It seems there was a problem with the edits/image pathg validation when
we were specifying uris in configuration. This patch ensures that the
uris are converted to names.

In the future, we will have to add some new logic for handling
bookkeeper or other shared devices.

Test Plan: amended test cases

Reviewers: hkuang, pritam, avf