Make sure HttpServer on the namenode shuts down quickly.
Summary:
The HttpServer on the namenode would take about 8 minutes to
shutdown if some of the handler threads of its threadpool are still running.
There is no need to wait for so long since all Http requests are immutable.
Also, I enhanced the threadpool such that all its threads are daemon
threads so that they do not hold up the JVM during shutdown

Test Plan:
Added a unit test that reproduces that HttpServer shutdown
problem.

Reviewers: hkuang, tomasz, sdong, weiyan

Reviewed By: tomasz