Enable restoring failed storage by default
Summary:
It seems that we do not use this feature at all anywhere now, as by
default the code does not attempt to restore storage directories. This
means that once a storage directory is failed, we can never restore it
without shutting down primary.

Test Plan:
ant test afftected tests
ant test - to see if we rely on the default "false" logic anywhere else

Reviewers: pritam, hkuang

Reviewed By: hkuang

Task ID: 1746342