Make the port range smaller for avatar tests
Summary:
After a few queries on internet, I found that:

/proc/sys/net/ipv4/ip_local_port_range

specifies the range of ports for local outgoing connections.
By default: 32768   65534
I checked a few machines on Jenkins, and this range is pretty much taken
in half (15000 or so) by connections to mcrouter or mcproxy). I think
this explains why taking the random port might be a regression if taken
from the full range. Before we were always starting from 10K, so we
would almost never go beyond 32K.

With this diff I decrease the max port to 32000.

I don't know what the dynamics of the MC connections is, but if they are
frequently changin, this might interfere with our tests.

Test Plan: ant test on jenkins

Reviewers: hkuang, pritam

Reviewed By: pritam