Make streaming job input key/value serialization configurable
Summary: We would like to be able to override the way keys/values are serialized in streaming jobs to be able to pass length-prefixed binary blobs or Thrift-serialized data structures to streaming mappers and reducers. The alternative would be to escape all tabs and newline characters and un-escape them on the other side.

Test Plan:
Unit tests
Running map-reduce jobs with TableInputFormat

Reviewers: nzhang, rvadali, aching, kranganathan

Reviewed By: rvadali

CC: hbase-eng@, mattwkelly

Task ID: 1735916