OfflineImageDecompressor tool
Summary:
This tool is for decompressing compressed image files, so we can check
corruptions. For as long as we can get through the header:
-layout
-namespace id
-numFiles
-genstamp
-txid (optional)

syntactically, we should be able to decompress the rest. Also this tool
is good for fixing image corruptions, better than customizing loading
and saving code.

Test Plan:
manually, tested dfsdev image, decompressed and parsed with
oiv

Reviewers: hkuang, pritam, weiyan

Reviewed By: hkuang