Introduce logic for max number of failed checkpoints
Summary:
We need to have a safeguard against continuous rolls of edit log on the
primary, as it can potentially create a huge number of edit log
segments. With this patch, I limit the number of such unsuccessful
checkpoints to 10, after which the standby exits.

I am not sure, is it better to throw a RuntimeException, or just
unconditionally exit the standby?

Test Plan: added test case

Reviewers: hkuang, pritam

Reviewed By: hkuang

Task ID: 1746342