Allow killing succeeded tasks from web UI
Summary:
Sometimes a task tracker with a finished mapper's output gets into a
state where the task tracker seems to be up and running, but fetching
the map output from the task tracker keeps failing. The reducer has to
try a lot of times before the map task is failed by the job tracker.
This change provides a way for the user to kill the finished map task.

Test Plan:
tested with a sleep job on a single node setup

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: aching

Reviewed By: aching

CC: jeanxu

Task ID: 1777845