Introduce retries for AvatarNode-ZK interaction
Summary:
This diff adds retries for some operations used during failover
-isZkEmpty
-getLastTxId
-registerAsPrimary

This is due to failures on dfstest.

-writeLastTxId and writeSessionTxid were alreasy retried, I just unified
all these calls.

It seems that now we retriy for some operations, for some we don'y. I
think it's better to do it uniformely. Also I moved it to an external
file so it's easier to keep track of interaction of AvatarNode and ZK

Test Plan: all avatar tests

Reviewers: hkuang, pritam

Reviewed By: pritam