Fix testcase TestAvatarStartup
Summary:
D591608 broke this test. Previously isPrimaryAlive() in AvatarNode, when
it was throwing NPE exception, it was handled by createAvatarNode, which
was retrhrowing it as IOException. With my change the NPE was simply
thrown. Te testcase expected IOException, but got NPE.

Also reduced the number of retries for TestAvatarStartup. Since most of
the tests expect failure, we don't want to retry 15 times. This cut down
the time for this test significantly.

Test Plan: ant test TestAvatarStartup

Reviewers: pritam, hkuang, avf

Reviewed By: pritam