Deletion with hardlinks doesn't work - fix
Summary:
Hardlinks do not handle directory deletions at all.
This is the first attemt to fix it.

Also, we need to remove this horrible function getTotalFiles or sth like
this. It's not updating the counts correctly when deleting directories.
Pritam, can we do sth about this?

Test Plan: added test, manually

Reviewers: hkuang, pritam, liyintang

Reviewed By: liyintang