Instrument the number of reducers which process no data(Hadoop changes)
Summary:
The aim of the task is to add the counters for the reducers which
process no data. A simple example can be when the hive query is given on an
empty partition of a table. I have added counters in ReduceTask to intrument
such reducers. The changes doesnt occur in open source hive. Note that these
counters are active only when the old reducer runs(and so are many other counters).

Test Plan:
I have tested that the counters for the reducers which process some
data are properly generated by running the sleep job on the local corona on my
dev box. Will also test these changes on the test cluster with hive queries on
empty tables so that the reducers with no data are also incremented.

Reviewers: kevinwilfong, njain, sambavim, rvadali

Reviewed By: rvadali

CC: mukundn

Task ID: 1750005