HDFS: fix TestFileAppend4
Summary:
TestFileAppend4.testNotPrematurelyCompleteWithFailure() appends a file and checks the pipeline can be recovered after one DN along the pipeline is killed. However, it is possible that only one replica is reported after close() so when the file is appened there is only one replica.

It is a design flaw that we might want to fix, but for now, fix the test to wait for all the three replicas to be reported before appending it.

Test Plan: N/A

Reviewers: tomasz, weiyan, hkuang

Reviewed By: weiyan