Enable buffered IO for log4j
Summary:
This patch enables bufferedIO for hadoop daemons, and sets the buffer
size to 4096, which should be slightly below 20 lines (the average size
in dfs1 is about 250bytes/line).

So obvoiusly we will not see synchronously those last 20 lines, but I
don't think it's a huge issue. On the other hand, my experiments show,
that with this settings we can speed up logging by a factor of 2.

I also added a hook which when the node is shutdown with hadoopctl
hadoop-daemon.sh stop - it will close the log manager which flushes the
contents of the buffer, so we won't miss those messages. Of course if we
kill a node with -9, we will loose up to those 20 lines.

-limited the classname to one from previous 2. The logs on dfs1 are
huge, we need
to make a pass I think over messages and limit these. In the last few
hours it's like 5GB per hour.

Test Plan: manually local setup, validated that log is flushed at normal shutdown

Reviewers: hkuang

Reviewed By: hkuang