Check the log disk free space
Summary:
The cluster manager can require a node to have a minimum amount of free
disk space before assigning any tasks on it. The corona task tracker
reports this free space as the maximum free space on any of its
configured disks. But we need to treat the log disk in special way. If
the log disk is full, we will not be able to launch any tasks even if
other disks have free space.
This diff reports the free disk space as Math.min(freeDiskSpace,
freeLogDiskSpace) to incorporate the special treatment of the log disk.

Test Plan:
Created a ram disk and specified that as the log directory. Verified
that the log disk space is obtained correctly through log messages.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: aching, jeanxu

Reviewed By: aching

CC: pknowles

Task ID: 1798718