Fix the cleanup of jtlogs
Summary:
The log cleanup works on a single level only - it lists files under a
single directory and cleans up the old files. The jtlogs directory was
not getting cleaned up because it has a single subdirectory under it -
userlogs - that contains all the actual logs. So it was not getting
purged. This fix passes the correct directory to the cleanup logic.

Test Plan:
created a file with old timestamp under jtlogs/userlogs and
verified that it got deleted.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: aching, jeanxu

Reviewed By: aching

CC: pknowles

Task ID: 1798718, 1466066