Remove getFilesTotal() from FSNamesystem
Summary:
The getFilesTotal() function was reporting incorrect counts and
was also unnecessary since the probability of choosing a file during
sampling is same as the percentage of files to sample.

Also, keeping a count of the total number of files and updating it each
time we change something else is error prone.

Test Plan: All unit tests.

Reviewers: tomasz, weiyan, hkuang

Reviewed By: tomasz

Task ID: 1801746