Add purging of ckpt images
Summary:
This patch adds purging of ckpt images:
-we purge only the ckpt images that are older than the oldest retained
good checkpoint, for safety.

Changed the location of purging, we cannot do this when upload finishes,
because the standby is still doing validation. We can safely do this
when the image is rolled correctly. Also The purging is placed outside
of the write lock, since it involves deleting images, which can be
potentially expensive.

Test Plan: added test

Reviewers: hkuang, pritam

Reviewed By: hkuang