Clean up checkpoint properly on primary/standby
Summary:
This diff:
-fixes the FSImage.deleteCancelledCheckpoint logic -> it was using
actual image name except the "ckpt version. It was wrong in a sense that
it wouldn't remove anything.

-call to remove cancelled checkpoint on standby quiesce

-found a bug in save namespace where edit log was unnecesarily opened -
it's opened in the finally clause one funcion above, I still need to run
unit tests, but this would affect unlikely situation of saving image on
the primary and cancelling this -> we normally cancell on standby.

Test Plan:
added tests to ensure that checkpoints are cleaned up after cancellation
on primary and standby

Reviewers: pritam, hkuang

Reviewed By: hkuang