HDFS: Fix bug of truncateBlock() when no partial chunk
Summary:
Currently, when truncateBlock(), the last chunksum is always recalculated using the last chunk. For the case that the last chunk is not partial, as a mistake, the last zero size chunk is used, polluting the data.
With this patch, the last checksum is not recalculated if last chunk is not partial.

This patch would fix TestFileAppend4.testAppendSyncReplication?()

Test Plan: ant test

Reviewers: hkuang, tomasz

Reviewed By: tomasz