Simplify standby backup logic
Summary:
I think that the best model for standby backup is when we keep up to
maximum number of backups, and we simply not allow for the the standby
to start if there is more backups. This will force manual intervention,
but will prevent the standby from starting up over and over again.

I think this is the best solution.

Test Plan:
added tests which ensure that the retentions manager is doing its job
properly

Reviewers: hkuang, pritam

Reviewed By: hkuang