Another source of mismatched generation stamp replicas
Summary: Looks that NN accepts a replica which has higher generation stamp than the one in memory if it is not the last block of an under-construction file. This diff changes the behavior to reject such a replica. Another option to make datanode to replicate and provision a replica with higher generation stamp.

Test Plan: TestCorruptBlocks

Reviewers: tomasz, sdong

Reviewed By: tomasz

Task ID: 1783644