Make the injections to be the last statements of doCheckpoint
Summary:
TestAvatarCheckpointingFailures failed because the testcase was notified
earlier than the checkpoint failed than handleCheckpointFailure function
was called. This diff makes the injections to be the last statements of
checkpoint to make sure all actions have already taken place when the
checkpoint notifies testing handlers that it completed.

Make the runtime volatile in FSEditLog as well to make sure it's not
cached across threads.

Test Plan: ant test TestAvatarCheckpointingFailures

Reviewers: hkuang, pritam

Reviewed By: hkuang