Online upgrade of task trackers
Summary:
To support online upgrade of TT, CoronaAdmin has 2 new commands:restartTaskTracker and forceRestartTaskTracker. CoronaAdmin will send
the requests to CM. If it is for restartTaskTracker, the current TTs will be added to a working list. If it is for forceRestartTaskTracker,
the current TTs will replace the working list. CM will flip the list of TTs to "ready" in batches. When a TT sends a heartbeat to CM, if it is
"ready", a restart reply will be sent back to the TT. The TT will exit.  After a restart request is sent out from CM, it will be removed from
the CM working list.

Test Plan: Test with sleep jobs.

Reviewers: rvadali, aching, pamelavagata

Reviewed By: rvadali

Task ID: 1741430