Backup Corona releases so that Corona job tracker will not be interrupteded by new relases
Summary:
1)	Corona task tracker will create a thread CoronaReleaseManager. This thread is responsible for managing
the releases.
2)	Any remote job will get the latest release from this manager and return the release to it after it is done.
3)	When the release manager starts, and for every run, it will check the release directory.s jar file/class files
timestamps, and make a copy (hardlink doesn.t work) if there is a new one
4)	Once all the files are copied, a flag file will be put into the copied directory. So if there is no change
between tasktracker restarts, there is no new copy made
5)	CoronaReleaseManager maintains a list of releases  and the jobids using them. It will check periodically if
there is any old release that no one will use it any more. If yes, it will remove the old release. The latest release
will stay even there is no one using it.
6)	CoronaReleaseManager will also scan the copied directories. If a directory is too old (more than 2 days.conf),
and if it is not in the internal list, it will be removed.

Test Plan: Test with the sleep job

Reviewers: rvadali, aching

Reviewed By: rvadali

Task ID: 1794271