Make storage directory failures consistent
Summary:
Now we have two levels of handling problems with journals.
-NNStorage level - what we previously had
-JournalAndStream - enabled/disabled functionality

With this diff, when we set a journal to be disabled, I also want to
report it to failed storage directories, if it's a file journal manager.

The orther direction is already handled, when we start new log segment,
we call attemtRestoreRemovedStorage and the directory is restored, then
the JournalAndStream is reinitialized with enabled option.

I added listing f journals and their state on web ui as well.

also unified some loging, so it's easier to grep

Test Plan: ant test, manually

Reviewers: hkuang, pritam, sdong

Reviewed By: hkuang