Fix problem with handling failed journals
Summary:
I discovered two problems when palying with dfstest2 today.

1) When a required journal fails on rollEditLog

JournalSet.mapJournalsAndReportErrors reports one failed journal to
disableAndReportErrorOnJournals, which in turn calls checkJournals(),
which throws IOException. Normally this exception is handled in
FSEditLog.sync, where we call runtime.exit. But here, since we roll the
edit log, the exception is just thrown to the caller, and the failed
journal is disables. Any subsequent operation (logEdit, logSync, etc)
will continue to write to the remaining journal. The nn will not exit.

With this diff, I set a flag that the exception is thorwn, and the next
time when we call disableAndReport... checkJournals() is called and
exception is rethrown, and will be handled in Sync(). I could call
checkJournals always, but I want to avoid doing this, and hence the use
of the flag.

2) When both journals fail on roll, both are eliminated, and then
editLogStream becomes null, and edit log operations fail on NPE, which
is also not good, because this will kill all rpc threads, and the nn
will be still running. I added check for NPE in logEdit, and an
exception in sync is thorwn to trigger shutdown.

Test Plan: added two test cases for the described situations, manually as well

Reviewers: hkuang, pritam, avf

Reviewed By: hkuang