Speed up offline image viewer
Summary:
After talking to Justin, it seems that this tool is used extensively,
and is very slow.
I looked into it, and I found a few inefficiencies

-Formatter.format- very expensive
-readin permissions
-parsing unnecessary ints and longs
-buffered writer

Altogether, I tested it on some old DFS1 image of size 7.3G

-before- ~35m 30s on my dev box
-after-  ~17m 00s

So we should get x2 speedup for this.

Test Plan: manually

Reviewers: hkuang, sdong

Reviewed By: hkuang