Add check to avoid NPE in FSNamesystem close.
Summary:
I ported this to alligator but I moved the check for policy to the
beginning of initialize, and then the "dir" object is not initialized.
If the policy check throws IOException, then FSNamesystem.close() is
called and dir.close() throws NPE.

In trunk that doesn't happen since dir is already initialized. But the
additional check won't do any harm, hece I'm submitting this on trunk.

Test Plan: ant test

Reviewers: pritam, junfang, dikang

Reviewed By: dikang