More optimizations for offline image viewer
Summary:
With this this I get rid of a bunch of unnecessary string coversions,
date conversions, and other calls for elements that are not printed in
the output.
This gives additional savings - I estimate 15+% - but more importantly
for standby - less heap pressure.

The shortest time that I got from this is 12 minutes for DFS1 image
(previously taking 35 minutes), although this might be a bit biased
because of os cache.

Test Plan: compare output

Reviewers: hkuang, pritam

Reviewed By: hkuang