Memory leak on standby's
Summary:
This diff adds cleaning of replication queues to standby. It turns out
that we did not ever see the sizes of most of these queues because the
sizes are updated in computeDataNodeWork(), which always exits because
of the safemode.

This is most likely the reason of the memory leak. At this moment on
dfs1. Just before restart on Saturday there were 6M+ missing blocks.
This is the count I can see, all other counts are hidden. They show to
be 0, but they arent...

Test Plan: ant test

Reviewers: hkuang, pritam, weiyan

Reviewed By: hkuang

Task ID: 1845279