Upgrade ZooKeeper to 3.4.3
Summary:
Upgrade ZooKeeper to 3.4.3 for bug fixes and compatibility with
BookKeeper.  This revision updates the ZooKeeper charge and modifies
MiniAvatarCluster to use new classnames for started the embedded
ZooKeeper server.

Test Plan:
Compile and run all unit tests for high-availability module Verify
that we can connect to and communicate with existing ZooKeeper
clusters that may use an older version of ZooKeeper.

Revert Plan:

Reviewers: hkuang, tomasz